Create callback_basic.py
Example of using callback to trigger a subsequent function once the action has completed.  in this case it will use the IP from the results of the geolocate ip action as the input for running "ip reputation" from the virus total integration. In addition it also dumps debug info for the various sections.